move transformSource into utils
