+++ packages/components/codemods/__tests__/parseJsx.ts
@@ -1,0 +1,11 @@
+import ts from "typescript"
+
+// This function is used to parse a stringified JSX element into an AST
+export const parseJsx = (jsx: string): ts.SourceFile =>
+  ts.createSourceFile(
+    "tempFile.tsx",
+    jsx,
+    ts.ScriptTarget.Latest,
+    true,
+    ts.ScriptKind.TSX
+  )
